address lookup, contract optimalization, and more!

+++ eth/contracts/MessageList.sol
@@ -5,24 +5,24 @@ contract MessageList {
     struct Message {
         string content;
         address senderAddr;
-        uint256 time;
+        uint time;
     }
 
     struct Client {
         bool initialized;
-        uint256 messageCount;
-        mapping(uint256 => Message) messages;
+        uint messageCount;
+        mapping(uint => Message) messages;
     }
 
     event MessageCreated(
-        uint256 count,
+        uint count,
         address sender,
         address reciever
     );
 
     mapping(address => Client) private clients;
 
-    function createMessage(string memory _content, address reciever) public returns (uint256) {
+    function createMessage(string memory _content, address reciever) public returns (uint) {
         Client storage recv = clients[reciever];
 
         if (!recv.initialized) {
@@ -47,14 +47,14 @@ contract MessageList {
 
         Message[] memory res = new Message[](recv.messageCount+sndr.messageCount);
         
-        uint256 j = 0;
-        for (uint256 i = 0; i<=recv.messageCount; i++ ) {
+        uint j = 0;
+        for (uint i = 0; i<=recv.messageCount; i++ ) {
             if(sender == recv.messages[i].senderAddr){
                 res[j] = recv.messages[i];
                 j++;
             }
         }
-        for (uint256 i = 0; i<=sndr.messageCount; i++ ) {
+        for (uint i = 0; i<=sndr.messageCount; i++ ) {
             if(reciever == sndr.messages[i].senderAddr){
                 res[j] = sndr.messages[i];
                 j++;
